requirements.txt
ultralytics>=8.0.0
opencv-python-headless>=4.8.0
pyyaml>=6.0
aiohttp>=3.9.0
websockets>=12.0
backoff>=2.2.1
packaging>=23.0
numpy>=1.24.0
